chore: move entity to package root to hinder circular imports
