+++ gliner_api/datamodel.py
@@ -1,5 +1,6 @@
-from pydantic import AliasChoices, BaseModel, Field, TypeAdapter
+from pydantic import BaseModel, Field, TypeAdapter
 
+from gliner_api import Entity
 from gliner_api.config import Config, get_config
 from gliner_api.examples import Examples, get_examples
 
@@ -10,29 +11,6 @@ examples: Examples = get_examples()
 class ErrorMessage(BaseModel):
     error: str = Field(description="Short error code")
     detail: str = Field(description="Detailed error explanaiton")
-
-
-class Entity(BaseModel):
-    start: int = Field(
-        ge=0,
-        description="Start index of the entity in the input text",
-    )
-    end: int = Field(
-        ge=0,
-        description="End index of the entity in the input text",
-    )
-    text: str = Field(
-        description="Text of the entity, extracted from the input text",
-    )
-    type: str = Field(
-        validation_alias=AliasChoices("type", "label"),
-        description="Entity type or label",
-    )
-    score: float = Field(
-        ge=0.0,
-        le=1.0,
-        description="Confidence score of the entity detection, between 0 and 1",
-    )
 
 
 class InvokeRequest(BaseModel):
